+++ kipper/core/src/compiler/tokens/expressions.ts
@@ -757,6 +757,11 @@ export interface FunctionCallPostfixExpressionSemantics extends ExpressionSemant
 	 * @since 0.5.0
 	 */
 	function: KipperFunction;
+  /**
+   * The arguments that were passed to this function.
+   * @since 0.6.0
+   */
+  args: Array<Expression<any>>;
 }
 
 /**
@@ -789,11 +794,12 @@ export class FunctionCallPostfixExpression extends Expression<FunctionCallPostfi
 		this.semanticData = {
 			evaluatedType: "void",
 			identifier: identifierSemantics.identifier,
-			// Tries to fetch the function. If it fails throw a {@link UnknownFunctionIdentifier} error.
+			args: this.children.slice(1, this.children.length), // Every item from index 0 to the end is an argument
+      // Tries to fetch the function. If it fails throw a {@link UnknownFunctionIdentifier} error.
 			function: this.programCtx.assert(this).getExistingFunction(identifierSemantics.identifier),
 		};
 
-		// Type checking
+		// TODO! Type checking
 	}
 
 	/**